+++ Packages/com.unity.inputsystem/InputSystem/Controls/Vector2Control.cs
@@ -36,8 +36,6 @@ namespace UnityEngine.InputSystem.Controls
         [InputControl(offset = 4, displayName = "Y")]
         public AxisControl y { get; set; }
 
-        public override float magnitude => value.magnitude;
-
         /// <summary>
         /// Default-initialize the control.
         /// </summary>
@@ -64,8 +62,8 @@ namespace UnityEngine.InputSystem.Controls
                     return *(Vector2*)((byte*)statePtr + (int)m_StateBlock.byteOffset);
                 default:
                     return new Vector2(
-                        x.ReadUnprocessedValueFromState(statePtr),
-                        y.ReadUnprocessedValueFromState(statePtr));
+                        x.ReadUnprocessedValueFromStateWithCaching(statePtr),
+                        y.ReadUnprocessedValueFromStateWithCaching(statePtr));
             }
         }
 
@@ -88,7 +86,7 @@ namespace UnityEngine.InputSystem.Controls
         public override unsafe float EvaluateMagnitude(void* statePtr)
         {
             ////REVIEW: this can go beyond 1; that okay?
-            return ReadValueFromState(statePtr).magnitude;
+            return ReadValueFromStateWithCaching(statePtr).magnitude;
         }
 
         protected override FourCC CalculateOptimizedControlDataType()