+++ Packages/com.unity.inputsystem/InputSystem/Controls/Vector3Control.cs
@@ -16,8 +16,6 @@ namespace UnityEngine.InputSystem.Controls
         [InputControl(offset = 8, displayName = "Z")]
         public AxisControl z { get; set; }
 
-        public override float magnitude => value.magnitude;
-
         public Vector3Control()
         {
             m_StateBlock.format = InputStateBlock.FormatVector3;
@@ -40,9 +38,9 @@ namespace UnityEngine.InputSystem.Controls
                     return *(Vector3*)((byte*)statePtr + (int)m_StateBlock.byteOffset);
                 default:
                     return new Vector3(
-                        x.ReadUnprocessedValueFromState(statePtr),
-                        y.ReadUnprocessedValueFromState(statePtr),
-                        z.ReadUnprocessedValueFromState(statePtr));
+                        x.ReadUnprocessedValueFromStateWithCaching(statePtr),
+                        y.ReadUnprocessedValueFromStateWithCaching(statePtr),
+                        z.ReadUnprocessedValueFromStateWithCaching(statePtr));
             }
         }
 
@@ -64,7 +62,7 @@ namespace UnityEngine.InputSystem.Controls
         public override unsafe float EvaluateMagnitude(void* statePtr)
         {
             ////REVIEW: this can go beyond 1; that okay?
-            return ReadValueFromState(statePtr).magnitude;
+            return ReadValueFromStateWithCaching(statePtr).magnitude;
         }
 
         protected override FourCC CalculateOptimizedControlDataType()